==> 01_training_testing_data.ipynb <==
# Training and Testing Data

A premlinary step in a machine learning project is to split the full data set into a *training* set and a *testing* set.  The training set will be used for feature selection, feature engineering, and hyperparameter tuning.  The testing is used only after the training process has been completed, and serves as a final sanity check that the model is working as expected on data that it hasn't been trained on.

The purpose of this notebook is to create the training set and testing set.

## Importing Pacakges

I am using the **polars** package because it can better handle large data sets.

import polars as pl

## Reading-In Full Data Set

df_full = pl.read_csv("accepted_2007_to_2018q4/accepted_2007_to_2018Q4.csv", ignore_errors=True)
df_full

## Dropping Rows with All `null` Values

I noticed visually that there are some rows that have all null values.  So let's remove those now.  It turns out that there are only a few dozen of those.

# Drop rows where all values are null
df_full = df_full.filter(~pl.all_horizontal(pl.all().is_null()))
df_full

## Shuffling Data

Next, we do a random shuffle of the data in case there is any inherent ordering to the CSV file.

df_full = df_full.sample(n=len(df_full), shuffle=True, seed=0)

## Columns to Retain

This is a curated list of features to start with that removes mostly null columns and also removes features that seem to be observed after a loan has been issued.

keep_list = ['loan_status','funded_amnt','addr_state', 'annual_inc', \
'application_type','dti', 'earliest_cr_line', 'emp_length',\
'emp_title', 'fico_range_high',\
'fico_range_low', 'grade', 'home_ownership', 'id', 'initial_list_status', \
'installment', 'int_rate', 'loan_amnt', \
'mort_acc', 'open_acc', 'pub_rec', 'pub_rec_bankruptcies', \
'purpose', 'revol_bal', 'revol_util', \
'sub_grade', 'term', 'title', 'total_acc',\
'verification_status', 'zip_code','last_pymnt_amnt',\
'num_actv_rev_tl', 'mo_sin_rcnt_rev_tl_op',\
'mo_sin_old_rev_tl_op',"bc_util","bc_open_to_buy",\
"avg_cur_bal","acc_open_past_24mths" ]

# Determining Training and Testing Set Size

I am going to use 70% of the data for training, and 30% of the data for testing.

0.7 * len(df_full)

size_training = 1582468
size_test = len(df_full) - size_training

## Writing Data to CSVs

df_full[0:1582468][keep_list].write_csv("data_training_testing/lending_club_training.csv")

df_full[1582468:][keep_list].write_csv("data_training_testing/lending_club_testing.csv")
