==> 02_binary.ipynb <==
# Isolate Binary Outcomes

The purpose of this notebook is to isolate the loans that have a `loan_status` value of `Fully Paid` or `Charged Off` so as to turn this into a binary classification problem.

### Load Packages

import pandas as pd
pd.options.display.max_rows = 200

### Read-In Data

df_training = pd.read_csv("data_training_testing/lending_club_training.csv")
df_testing = pd.read_csv("data_training_testing/lending_club_testing.csv")

df_training.head(1).T

### Explore Data

df_training['loan_status'].value_counts()

df_testing['loan_status'].value_counts()

### Separate Out Binary Outcomes

We are going to ignore all the loans that aren't `Fully Paid` or `Charged Off`.

df_binary_training = df_training.query('loan_status == "Fully Paid" | loan_status == "Charged Off"').copy()
df_binary_testing = df_testing.query('loan_status == "Fully Paid" | loan_status == "Charged Off"').copy()

### Encoding Loan Status

Let's create a new label column called `charged_off` which is the the one-hot encoding of `loan_status`.

df_binary_training['charged_off'] = pd.get_dummies(df_binary_training[['loan_status']])['loan_status_Charged Off']
df_binary_training.drop(columns=['loan_status'], inplace=True)

df_binary_testing['charged_off'] = pd.get_dummies(df_binary_testing[['loan_status']])['loan_status_Charged Off']
df_binary_testing.drop(columns=['loan_status'], inplace=True)

### Write to CSV

df_binary_training.to_csv('data_processed/01_binary_training.csv', index=False)
df_binary_testing.to_csv('data_processed/02_binary_testing.csv', index=False)
