==> 03a_pipeline_model_fits_categorical.ipynb <==
## Import Packages

import pandas as pd
import numpy as np

## Read-In Data

df_default = pd.read_csv("data_processed/01_binary_training.csv", low_memory=False)

## Features List

features_categorical = [
    #'earliest_cr_line', # need to handle this in a pipeline
    'grade',
    'home_ownership',
    'purpose',
    'sub_grade',
    'term',
    'verification_status',
]

from sklearn.compose import ColumnTransformer
from sklearn.impute import SimpleImputer
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import OrdinalEncoder
from sklearn.pipeline import Pipeline

categorical_transformer_1 = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='most_frequent')),
    ('ordinal', OrdinalEncoder(categories=[['A', 'B', 'C', 'D', 'E', 'F', 'G']]))
])

model = Pipeline([
    ("select", 
        ColumnTransformer(
            [("keep", "passthrough", ["grade"])],
            remainder="drop"
        )
    ),  # keeps certain columns and drops everything else
    ("preprocessor", categorical_transformer_1),
    ("model", LogisticRegression())
])

X = df_default.drop(columns=(['charged_off']))
y = df_default['charged_off']

model.fit(X, y)

model.score(X, y)

model.predict(X).mean()

default_balance_actual = (df_default["charged_off"] * df_default["loan_amnt"]).sum()
default_balance_actual

default_balance_hard_prediction = (model.predict(X) * df_default["loan_amnt"]).sum()
default_balance_hard_prediction

default_balance_hard_prediction / default_balance_actual

default_balance_expected = (model.predict_proba(X)[:,1] * df_default["loan_amnt"]).sum()
default_balance_expected

default_balance_expected / default_balance_actual

## Threshold Tuning

# predict probabilities
yhat = model.predict_proba(X)

# keep probabilities for the positive outcome only
probs = yhat[:, 1]
probs

# define thresholds
thresholds = np.arange(0, 1, 0.01)

# apply threshold to positive probabilities to create inferences
def to_inference(pos_probs, threshold):
	return (pos_probs >= threshold).astype('int')

from sklearn.metrics import f1_score, accuracy_score
# evaluate each threshold
scores = [accuracy_score(y, to_inference(probs, t)) for t in thresholds]

# get best threshold
ix = np.argmax(scores)
print('Threshold=%.3f, Score=%.5f' % (thresholds[ix], scores[ix]))

(to_inference(probs, 0.64) * df_default["loan_amnt"]).sum()

## Random Forest Classifier

from sklearn.ensemble import RandomForestClassifier

model = Pipeline([
    ("select", 
        ColumnTransformer(
            [("keep", "passthrough", ["grade"])],
            remainder="drop"
        )
    ),  # keeps certain columns and drops everything else
    ("preprocessor", categorical_transformer_1),
    ("model", RandomForestClassifier(n_estimators=50, max_depth=10, n_jobs=-1))
])

model.fit(X, y)

model.score(X, y)

model.predict(X).mean()

default_balance_expected = (model.predict_proba(X)[:,1] * df_default["loan_amnt"]).sum()
default_balance_expected

default_balance_actual = (df_default["charged_off"] * df_default["loan_amnt"]).sum()
default_balance_actual

default_balance_expected / default_balance_actual

# predict probabilities
yhat = model.predict_proba(X)

# keep probabilities for the positive outcome only
probs = yhat[:, 1]

# define thresholds
thresholds = np.arange(0, 1, 0.01)

from sklearn.metrics import f1_score, accuracy_score
# evaluate each threshold
scores = [f1_score(y, to_inference(probs, t)) for t in thresholds]

# get best threshold
ix = np.argmax(scores)
print('Threshold=%.3f, Score=%.5f' % (thresholds[ix], scores[ix]))

(to_inference(probs, 0.14) * df_default["loan_amnt"]).sum()

probs.mean()

y.mean()
