==> 06_threshold_tuning.ipynb <==
# Threshold Tuning and Saving to Pickle File

In this notebook we will perfrom threshold tuning and save our model to a pickle file.

## Importing Packages

import pandas as pd
import numpy as np

## Reading-In Training Data

df_default = pd.read_csv("data_processed/01_binary_training.csv", low_memory=False)
df_default.head().T

## Feature Selection

Moving forward we will focus on the numeric features only.

numeric_features = [
    "funded_amnt",
    "last_pymnt_amnt",
    "int_rate",
    "loan_amnt",
    "installment",
    "acc_open_past_24mths",
    "dti",
    "fico_range_low",
    "mort_acc",
]

## Creating the `FeatureSelector` Column Transformer

Creating the customer feature selector which will be the first step in all the pipelines.

from sklearn.base import BaseEstimator, TransformerMixin

class FeatureSelector(BaseEstimator, TransformerMixin):
    def __init__(self, columns):
        self.columns = columns
        
    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        return X[self.columns]

## Preprocessing

Next we use `Pipelines` and `ColumnTransformers` to affect preprocessing of our data.

from sklearn.impute import SimpleImputer
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.ensemble import RandomForestClassifier

numerical_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='mean')),
    ('scaler', StandardScaler(with_mean=True))  
])

preprocessor = ColumnTransformer(transformers=[
    ('numerical', numerical_transformer, numeric_features),],
    remainder = 'passthrough',
)

## Model Definition and Fitting

Now, lets' put our preprocessor together with a `RandomForestClassifier` estimator and fit our model.

model = Pipeline(steps=[
    ('feature_selector', FeatureSelector(numeric_features)),
    ('preprocessor', preprocessor),
    ('random_forest', RandomForestClassifier(n_estimators=50, max_depth=10, n_jobs=-1, random_state=0))
])

X = df_default.drop(columns=(['charged_off']))
y = df_default['charged_off']

model.fit(X, y)

As we can see, our model has decent accuracy and predicts a reasonable number of defaults.

model.score(X, y)

model.predict(X).mean()

## Threshold Tuning

The default threshold of discrimination for **sklearn** classifiers is 0.5.  Sometimes with imbalanced data sets performance can be improve by choosing a different threshold.  

We will now search for the optimal threshold.

# predict probabilities
yhat = model.predict_proba(X)

# keep probabilities for the positive outcome only
probs = yhat[:, 1]
probs

# define thresholds
thresholds = np.arange(0, 1, 0.01)

# apply threshold to positive probabilities to create inferences
def to_inference(pos_probs, threshold):
	return (pos_probs >= threshold).astype('int')

from sklearn.metrics import f1_score, accuracy_score
# evaluate each threshold
scores = [accuracy_score(y, to_inference(probs, t)) for t in thresholds]

It looks like the optimal threshold for our model is 0.47, so fairly close to 0.5.

# get best threshold
ix = np.argmax(scores)
print('Threshold=%.3f, Accuracy=%.5f' % (thresholds[ix], scores[ix]))

The overall number of defaults is slightly improved with this threshold.

to_inference(probs, 0.48).mean()

## Saving Fitted Model to Pickle

Finally, let's save our fitted model to a pickle file.

import joblib

joblib.dump(model, "pickle/lending_club_model.pkl")
