==> 07_testing_the_model.ipynb <==
# Testing the Model

Recall that at the outset of this modeling exercise, we separated out a test set to evaluate our final model.  We are now ready to do so. 

## Importing Packages

import joblib
import pandas as pd

## Feature Selector

from sklearn.base import BaseEstimator, TransformerMixin

class FeatureSelector(BaseEstimator, TransformerMixin):
    def __init__(self, columns):
        self.columns = columns
        
    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        return X[self.columns]

## Reading-In Our Fitted Model Object

model_saved = joblib.load("pickle/lending_club_model.pkl")

## Reading-In Our Test Set

df_test = pd.read_csv("data_processed/02_binary_testing.csv")
df_test

## Model Accuracy with 0.5 Threshold

Notice that we have to do no preprocessing to our test data.  All the preprocessing is built into the pickle file.

from sklearn.metrics import f1_score, accuracy_score
print(accuracy_score(model_saved.predict(df_test), df_test["charged_off"]))
print(f1_score(model_saved.predict(df_test), df_test["charged_off"]))

## Using Optimal Threshold

Recall from our previous work thath the optimal threshold is 0.48.  Let see how that threshold affects accuracy.

# apply threshold to positive probabilities to create labels
def to_inference(pos_probs, threshold):
	return (pos_probs >= threshold).astype('int')

Accuracy is very slighly improved.

accuracy_score(df_test["charged_off"], to_inference(model_saved.predict_proba(df_test)[:,1], 0.47))
